level-up learnset expansion

+++ include/pokemon.h
@@ -649,7 +649,7 @@ s8 Pokemon_GetFlavorAffinityOf(u32 monPersonality, int flavor);
  * @param[out] monLevelUpMoveIDs Pointer to a u16 array for storing the level up moves
  * @return int
  */
-int Pokemon_LoadLevelUpMoveIdsOf(int monSpecies, int monForm, u16 *monLevelUpMoveIDs);
+int Pokemon_LoadLevelUpMoveIdsOf(int monSpecies, int monForm, u32 *monLevelUpMoveIDs);
 
 void Pokemon_ApplyPokerus(Party *party);
 u8 Pokemon_HasPokerus(Party *party, u8 param1);
@@ -773,7 +773,7 @@ BOOL Pokemon_SetRotomForm(Pokemon *mon, int monForm, int moveSlot);
  * @param monForm
  * @param[out] monLevelUpMoves Pointer to a u16 array to store the move table
  */
-void Pokemon_LoadLevelUpMovesOf(int monSpecies, int monForm, u16 *monLevelUpMoves);
+void Pokemon_LoadLevelUpMovesOf(int monSpecies, int monForm, u32 *monLevelUpMoves);
 
 /**
  * @brief Play a Pokemon's cry, according to the given species and form number.
